dialogue/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
python-multipart

# 음성(STT/TTS)이 필요하다면 아래 항목 포함
azure-cognitiveservices-speech
